+++ tests/src/io/segment/android/info/test/InfoManagerTest.java
@@ -1,5 +1,6 @@
 package io.segment.android.info.test;
 
+import io.segment.android.Options;
 import io.segment.android.info.InfoManager;
 
 import org.json.JSONObject;
@@ -9,11 +10,10 @@ import org.junit.Test;
 import android.test.AndroidTestCase;
 
 public class InfoManagerTest extends AndroidTestCase {
-
-	private InfoManager manager = new InfoManager();
 	
 	@Test
 	public void testGet() {
+		InfoManager manager = new InfoManager(new Options());
 		JSONObject object = manager.build(this.getContext());
 		Assert.assertTrue(object.length() > 0);
 	}